requirements.txt
astroid==2.3.2
autopep8==1.4.4
boardgamegeek2==1.0.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.22.0
requests-cache==0.5.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
